model/requirements.txt
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
git+https://www.github.com/keras-team/keras-contrib.git
matplotlib
keras
tensorflow
keras-resnet
scipy
pillow
numpy
pandas_ml
pandas
sklearn
seaborn
scikit-image